7_task_agent/main.py
/**
 * @generated-by AI: matthewmli
 * @generated-date 2025-03-30
 */
import sys
from pathlib import Path

# 添加项目根目录到 sys.path
sys.path.insert(0, str(Path(__file__).parent))

from tools import (
    ToolRegistry,
    CalculatorTool,
    FileReadTool,
    FileWriteTool,
    WebSearchTool,
    WebFetchTool
)
from agent import TaskAgent
from dotenv import load_dotenv

load_dotenv()


def setup_tools() -> ToolRegistry:
    """
    设置工具注册表
    
    Returns:
        ToolRegistry: 工具注册表
        
    TODO:
    1. 创建 ToolRegistry 实例
    2. 注册所有工具：
       - CalculatorTool
       - FileReadTool
       - FileWriteTool
       - WebSearchTool
       - WebFetchTool
    3. 返回注册表
    """
    pass


def setup_agent() -> TaskAgent:
    """
    设置 Task Agent
    
    Returns:
        TaskAgent: Task Agent 实例
        
    TODO:
    1. 获取工具注册表
    2. 定义 Agent 的系统提示词
    3. 创建 TaskAgent 实例
    4. 返回 Agent
    """
    pass


def print_welcome():
    """
    打印欢迎信息
    
    TODO: 打印欢迎信息和命令列表
    """
    print("=" * 60)
    print("Task Agent - 任务执行助手")
    print("=" * 60)
    print("可用命令:")
    print("  - 输入任务描述，Agent 将执行任务")
    print("  - status: 查看任务状态")
    print("  - pause: 暂停任务")
    print("  - resume: 恢复任务")
    print("  - cancel: 取消任务")
    print("  - help: 显示帮助信息")
    print("  - exit: 退出程序")
    print("=" * 60)


def main():
    """
    主函数
    
    TODO:
    1. 设置 Agent
    2. 打印欢迎信息
    3. 进入命令行循环：
       a. 读取用户输入
       b. 根据输入执行相应操作：
          - status: 显示任务状态
          - pause: 暂停任务
          - resume: 恢复任务
          - cancel: 取消任务
          - help: 显示帮助
          - exit: 退出
          - 其他: 作为任务执行
       c. 显示执行结果
    """
    pass


def handle_special_command(agent: TaskAgent, command: str) -> bool:
    """
    处理特殊命令
    
    Args:
        agent: Task Agent 实例
        command: 命令字符串
        
    Returns:
        bool: 是否继续运行
        
    TODO:
    处理特殊命令：
    - status: 显示任务状态
    - pause: 暂停任务
    - resume: 恢复任务
    - cancel: 取消任务
    - help: 显示帮助
    - exit: 返回 False（停止运行）
    """
    pass


def print_task_result(result):
    """
    打印任务结果
    
    Args:
        result: 任务结果
        
    TODO:
    格式化显示任务结果：
    - 成功/失败状态
    - 消息内容
    - 详细信息（如果有）
    """
    pass


if __name__ == "__main__":
    main()
