7_task_agent/requirements.txt
# LLM 相关
openai>=1.0.0
python-dotenv>=1.0.0

# 数据模型
pydantic>=2.0.0

# Web 请求
requests>=2.31.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
